beta distribution for gpu pipeline complete
- need more testing
- need integration with the rest of the server

+++ tmb-server/stats_verify.h
@@ -1,8 +1,8 @@
 // for each row of the csv file
 typedef struct {
-    float gpuUtilization;
-    float vramUsage;
-    float powerDraw;
+    double gpuUtilization;
+    double vramUsage;
+    double powerDraw;
 } DataRow;
 
 // contains the entire CSV
